[IL-634] add todos

+++ contracts/assetManagment/interfaces/compound/CEth.sol
@@ -1,9 +1,14 @@
 pragma solidity ^0.8.0;
 
+// TODO: If not needed please remove
 interface CEth {
     function mint() external payable;
+
     function exchangeRateCurrent() external returns (uint256);
+
     function supplyRatePerBlock() external returns (uint256);
-    function redeem(uint) external returns (uint);
-    function redeemUnderlying(uint) external returns (uint);
+
+    function redeem(uint256) external returns (uint256);
+
+    function redeemUnderlying(uint256) external returns (uint256);
 }
